enhancement: form validation for authentication
