+++ src/components/common/input/index.tsx
@@ -8,8 +8,10 @@ import { cn } from "~/lib/utils";
 const inputVariants = cva("text-sm rounded-md transition-colors", {
   variants: {
     variant: {
-      primary: "border-primary text-primary focus:outline-none focus:border-primary",
-      border: "border-border text-foreground focus:outline-none focus:border-border",
+      primary:
+        "border-primary text-primary focus:outline-none focus:border-primary",
+      border:
+        "border-border text-foreground focus:outline-none focus:border-border",
     },
     state: {
       default: "border-border",
@@ -88,12 +90,12 @@ const Input = React.forwardRef<HTMLInputElement, InputProperties>(
             {label}
           </label>
         )}
-        <div className="flex items-center w-full">
+        <div className="flex w-full items-center">
           <Comp
             className={cn(
               inputVariants({ variant, state }),
               "w-full border px-4 py-2",
-              "flex h-10 rounded-md border-input bg-background text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:border-primary disabled:cursor-not-allowed disabled:opacity-50",
+              "flex h-10 rounded-md border-input bg-background text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:border-primary focus-visible:outline-none disabled:cursor-not-allowed disabled:opacity-50",
               className,
             )}
             ref={reference}
