+++ src/schemas/index.ts
@@ -1,18 +1,19 @@
 import * as z from "zod";
 
-const passwordSchema = z.string()
+const passwordSchema = z
+  .string()
   .min(1, { message: "Password is required" })
   .min(8, { message: "Password must be at least 8 characters long" })
-  .refine(value => /[a-z]/.test(value), {
+  .refine((value) => /[a-z]/.test(value), {
     message: "Password must contain at least one lowercase letter",
   })
-  .refine(value => /[A-Z]/.test(value), {
+  .refine((value) => /[A-Z]/.test(value), {
     message: "Password must contain at least one uppercase letter",
   })
-  .refine(value => /\d/.test(value), {
+  .refine((value) => /\d/.test(value), {
     message: "Password must contain at least one number",
   })
-  .refine(value => /[\W_]/.test(value), {
+  .refine((value) => /[\W_]/.test(value), {
     message: "Password must contain at least one special character",
   });
